requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.2.1
SQLAlchemy==2.0.30
alembic==1.13.1
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
boto3==1.34.162
botocore==1.34.162
